decoder: always return NANOCBOR_OK instead of bytes read
For now, number of bytes read are returned for all number-related functions, but not for containers, tags and strings.

This is an API breaking change. However, we did not come up with a good reason why users of NanoCBOR would be interested in the number of bytes read, especially if they only get this information for part of the API.